Update lbs_project.js to v4
Had to read up on measuring time. hrtime is built in to Node.js and has a high resolution (nano seconds). The values are extracted from an array where seconds are multiplied by 1000, and nanoseconds are divided by 1 million, then added together. This results in milliseconds being presented, as per Node's documentation.