Added or operator to change falsy values of "encrypted_distance" to 0
There appears to be no actual falsy values, the only thing causing two falsy values in a data set of 10 000 data points is that iteration 3382 and 4495 generate the exact same values for client A and client B. The encrypted computation does not result in the numerical value of 0 as the plaintext computation does. However, since the cause is that the SEAL-library can't handle the computation, although it should be 0, and no other issues exist, it should be safe to treat these instances of falsy values to equal 0.